Draft of ROUGE.
ROUGE appears to be so slow as to be unusable.

I think I will switch to a reimplementation instead
of the official script because of that. (It also
errors when calling additional functions, like
sentence splitting. This seems to be because the
package is so old the dependencies are out-of-date.)

Anyway, as part of integrating for the original
ROUGE's specific and annoying input directory
requirements and naming conventions, this commit
also rewrites the temp file saving / removal,
pulls it out of BLEU into a general utility
module, and changes most of the types around.